host C化 默认值以及缺少tensorinfo信息修复
# message auto-generated for no-merge-commit merge:
merge main into master

host C化 默认值以及缺少tensorinfo信息修复

Created-by: msprofDeveloper

Match-id-fdccaeb91bc0db59ca2e2b8e6434b77502b3da96

+++ analysis/csrc/association/cann/tree_analyzer.h
@@ -129,7 +129,7 @@ private:
 
     void UpdateComputeDescForFftsSituation(ComputeOpDescs &descs, const std::shared_ptr<Event> &track);
     void UpdateComputeDescForHcclSituation(ComputeOpDescs &descs, const std::shared_ptr<Event> &track,
-                                           uint32_t item_id);
+                                           uint64_t item_id);
 
 private:
     // 树的root节点
